chore: fix DEFAULT_OPERATION_SLOTS exports
