+++ src/composables/useTheme.ts
@@ -7,7 +7,6 @@ import { generateCodeSample } from '../lib/codeSamples/generateCodeSample'
 import { deepUnref } from '../lib/deepUnref'
 import { locales } from '../locales'
 import type { OperationSlot } from '../types'
-import { DEFAULT_OPERATION_SLOTS } from '../index'
 
 export interface ThemeConfig {
   highlighterTheme: {
@@ -169,6 +168,20 @@ interface LanguageConfig {
 }
 
 type GeneratorFunction = (lang: string, request: IOARequest) => string
+
+export const DEFAULT_OPERATION_SLOTS: OperationSlot[] = [
+  'header',
+  'path',
+  'description',
+  'security',
+  'parameters',
+  'request-body',
+  'responses',
+  'try-it',
+  'code-samples',
+  'branding',
+  'footer',
+]
 
 const availableLanguages: LanguageConfig[] = [
   {
